Move `CollateralOutputs` to `Cardano.Read.Ledger.Tx.CollateralOutputs`
