+++ lib/read/cardano-wallet-read.cabal
@@ -51,6 +51,7 @@ library
     Cardano.Read.Ledger.Tx.Certificates
     Cardano.Read.Ledger.Tx.CBOR
     Cardano.Read.Ledger.Tx.CollateralInputs
+    Cardano.Read.Ledger.Tx.CollateralOutputs
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -72,7 +73,6 @@ library
     Cardano.Wallet.Read.Eras.KnownEras
     Cardano.Wallet.Read.Hash
     Cardano.Wallet.Read.Tx
-    Cardano.Wallet.Read.Tx.CollateralOutputs
     Cardano.Wallet.Read.Tx.Eras
     Cardano.Wallet.Read.Tx.ExtraSigs
     Cardano.Wallet.Read.Tx.Fee
